% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vegtype_pie.R
\name{vegtype_pie}
\alias{vegtype_pie}
\title{Plotly piechart of vegetation types for geometric structure}
\usage{
vegtype_pie(df, pieVals)
}
\arguments{
\item{df}{Filtered df using filter_piechart}

\item{pieVals}{Column of percent values for vegetation types}
}
\value{
A plotly piechart
}
\description{
Plotly piechart of vegetation types for geometric structure
}
